refactor : changed price from int to Long

+++ src/main/resources/db/migration/V1__innit.sql
@@ -74,7 +74,7 @@ CREATE TABLE event_ticket_prices (
     id BIGINT AUTO_INCREMENT PRIMARY KEY,
     event_id BIGINT NOT NULL,
     ticket_type_id BIGINT NOT NULL,
-    price INT NOT NULL,
+    price BIGINT NOT NULL,
     quantity INT NOT NULL,
 
     FOREIGN KEY (event_id) REFERENCES  events(id),
